requirements.txt
python-telegram-bot==20.7
google-auth==2.17.3
google-auth-oauthlib==0.7.1
google-api-python-client==2.88.0
